マイグレーションファイルの修正と本番DBの整合性を保った修正

+++ src/main/resources/db/migration/V6__Add_ai_api_usage_limits.sql
@@ -1,7 +1,7 @@
 -- AI API利用制限テーブル作成（将来の拡張を考慮した汎用名）
 CREATE TABLE ai_api_usage_limits (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    user_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    user_id INT NOT NULL,
     provider VARCHAR(50) NOT NULL DEFAULT 'gemini',
     date DATE NOT NULL,
     daily_count INT NOT NULL DEFAULT 0,
